Simplify and improve writeChannelFromHandle
The old version called hFileSize, which is only valid for handles with
an associated file size -- but then did nothing with the return value.
Removed that call.

Also made the main driver tail-recursive, and removed the return value
from 'send' (which is always equal to the total size to write)