Fix getHostKey
The C function libssh2_session_hostkey returns a const char* where the
first byte is (often) a NULL byte. This causes the Haskell FFI to return
an empty String. Hence, we create a new FFI to libssh2_session_hostkey
that returns a Ptr CChar, that we then wrap in a function that returns a
base64 encoded String. This way we can capture the host key, including its
NULL byte, in a proper Haskell type.

Although this is a bug fix, this changes Haskell type signatures of
exported functions.

See #66.